requirements.txt
anyio==4.6.2.post1
APScheduler==3.6.3
attrs==24.2.0
beautifulsoup4==4.12.3
cachetools==4.2.2
certifi==2024.8.30
charset-normalizer==3.4.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
outcome==1.3.0.post0
packaging==24.2
PySocks==1.7.1
python-dotenv==1.0.1
python-telegram-bot==13.7
pytz==2024.2
requests==2.32.3
selenium==4.27.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tornado==6.4.2
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
